requirements.txt
streamlit==1.32.0
librosa==0.10.2
numpy==1.23.5
tensorflow==2.12.0
pillow==10.4.0
